Issue #19 Catch JSON decoding error in Start handler
The bug occurs due to the buffer size being too small for the reponse of
Start and the unhandled exception from JSON  decoding of the  incomplete
response from the buffer

This only partially fixes the bug. as in, it won't show the unhandled exception
error window instead it'll display the error that occured during the  decoding to
the user in form of a message box.

The next commit fixes the bug completely by increasing the size of the response buffer