Fix pseudostate in statusform
This is to fix:

HTTP/1.0" 500 47
ERRO Cannot load machine: machine crc does not exist
\\.\pipe\crc-http - - [29/Jun/2021:14:19:39 +0800] "GET /api/status HTTP/1.0" 500 47
ERRO Cannot load machine: machine crc does not exist
\\.\pipe\crc-http - - [29/Jun/2021:14:19:44 +0800] "GET /api/status HTTP/1.0" 500 47

in which the daemon returns a 500 (n result is returned), but we do need
to have some status shown. In the application we use a the Success flag
to determine if a call was successful. In this case, the StatusForm can
show a psuedostate of Stopped instead.

Without this change:

     Invoke(c, new StatusResult { CrcStatus = InitialState, OpenshiftStatus = "Stopped"});

will cause the StatusForm to flicker Stopped, Running on failing
requests that either timedout or were cancelled.